BitfieldField:
- ast: nodes: changing beginBit_ and endBit_ to signed integers so we can more easily detect out of range fields (0 is a valid first field).
- parser: states: code changes to check the bit field value is within the range of the declared underlying type.

+++ swizzle/parser/detail/ExtractBitValueFromToken.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <cstddef>
+#include <cstdint>
 #include <boost/utility/string_view.hpp>
 
 namespace swizzle { namespace lexer {
@@ -10,5 +10,5 @@ namespace swizzle { namespace parser { namespace detail {
 
     // @return the numeric value stored in @info, perform bounds checking against @underlying raise an exception
     // if the value in @info is out-of-bounds
-    std::size_t extractBitValueFromToken(const boost::string_view& underlying, const lexer::TokenInfo& info);
+    std::intmax_t extractBitValueFromToken(const boost::string_view& underlying, const lexer::TokenInfo& info);
 }}}
